reorganize the admin code a bit
